update get notice score, update search...

+++ search/src/main/java/campuslifecenter/search/repository/NoticeRepository.java
@@ -3,5 +3,5 @@ package campuslifecenter.search.repository;
 import campuslifecenter.search.model.NoticeSearch;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface NoticeRepository extends ElasticsearchRepository<NoticeSearch, Integer> {
+public interface NoticeRepository extends ElasticsearchRepository<NoticeSearch, Long> {
 }
